Refactor emailService to TypeScript
